Slow clock after boot
This experiment slows the system clock to 16MHz at the very
end of `imix::main::reset_handler()`.

Continued functioning of console output is ensured by alerting
the USART3 device to the change of clock speed.  The USART
implementation reconfigures its baud-rate register when it
receives this alert.

Expect to see glitches in the serial output when the clock changes.